+++ minMysql/sqlRow.h
@@ -44,7 +44,7 @@ class sqlRow : public QMapV2<QByteArray, QByteArray> {
 	 */
 	template <isEnum T>
 	[[nodiscard]] T rqe(const QByteArray& key) const {
-		int temp = 0;
+		i64 temp = 0;
 		rq(key, temp);
 		T t2 = T(temp);
 		return t2;
